requirements.txt
flask==3.0.0
streamlit==1.34.0
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.23
google-generativeai==0.7.2

chromadb==0.5.0

pypdf==4.0.1
pdfplumber==0.11.0

python-dotenv==1.0.1

numpy==1.26.4
requests==2.32.3
tqdm==4.66.
langchain-google-genai==1.0.7
googlesearch-python==1.2.3
beautifulsoup4==4.12.3
